// clients/Client3/assets/src/models/playerPrefs/RecordPropertys.ts
/**
 * 记录属性KEY
 */
export enum RecordPropertys {
    /**
     * 序号索引
     */
    INDEX = "index",
    /**
     * 总音量
     */
    VOLUME = "volume",
    /**
     * 音乐音量
     */
    MUSIC_VOLUME = "music_volume",
    /**
     * 音效音量
     */
    SOUND_VOLUME = "sound_volume",
    /**
     * 难度
     */
    DIFFICULTY = "difficulty",
    /**
     * 游戏记录列表
     */
    GAMES = "games",
    /**
     * 关卡记录
     */
    LEVELS="levels",
    /**
     * 星级
     */
    STARS="stars",
}